Fix #62, make entity ID default to 32 bit
Entity IDs should be larger by default for real-world applicability, as this also
restricts what CFDP can receive, not just what it sends.  Note CFDP will only use
the number of bytes required to express the value, so values less than 256 will
still only use 1 byte, regardless of this config.  This just allows use of larger
values.

+++ fsw/platform_inc/cf_platform_cfg.h
@@ -43,7 +43,7 @@
  ** \par Limits
  **         Must be one of uint8, uint16, uint32, uint64.
  */
-typedef uint8 CF_EntityId_t;
+typedef uint32 CF_EntityId_t;
 
 /**
  ** \cfcfg transaction sequence number size
